back_end/requirements.txt
Django==3.0.8
djangorestframework
djangorestframework-simplejwt==4.4.0
django-model-utils==4.0.0
django-cors-headers
pymysql
pandas
openpyxl